+++ packages/chaingraph-trpc/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @badaitech/chaingraph-trpc
 
+## 0.4.1
+
+### Patch Changes
+
+- feat: enhance MCPToolCallNode input handling by wrapping input keys in arguments object when missing
+
 ## 0.4.0
 
 ### Minor Changes